createdMakefile

+++ ft_strchr.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-char *ft_strchr(const char *str, size_t c)
+char *ft_strchr(const char *str, int c)
 {
     int i;
